+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- Added `set(FutureOr<T>)` method to `AsyncNotifier` for synchronous
+  initialization/update of `AsyncSnapshot` without a `SynchronousFuture`.
+- Decreased the minimum Dart SDK version to `3.0.0`.
+
 ## 1.0.1
 
 - Documentation updates.